gdp-dashboard/bigquery_requirements.txt
google-cloud-bigquery==3.13.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pandas==2.1.4
numpy==1.24.3
streamlit==1.28.2
plotly==5.17.0
openpyxl==3.1.2
